Python/requirements.txt
astroid==2.0.4
CacheControl==0.12.5
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
Faker==0.9.0
firebase-admin==2.13.0
google-api-core==1.3.0
google-auth==1.5.1
google-cloud==0.34.0
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.14.1
idna==2.7
isort==4.3.4
keyboard==0.13.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==2.2.3
mccabe==0.6.1
msgpack==0.5.6
numpy==1.15.1
pandas==0.23.4
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.1.1
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rsa==3.4.2
scipy==1.1.0
six==1.11.0
text-unidecode==1.2
tqdm==4.25.0
urllib3==1.23
wrapt==1.10.11
